# gateway: do not compute snihosts for service repeatedly