@ramaraochavali 2020-07-13
Before and After in benchmark

```
name                                   old time/op        new time/op        delta
ListenerGeneration/gateways-12                142ms ±12%         129ms ±11%    ~     (p=0.222 n=5+5)
ListenerGeneration/gateways-shared-12         306µs ± 9%         299µs ± 8%    ~     (p=0.690 n=5+5)
ListenerGeneration/empty-12                  9.88ms ±12%        9.45ms ±11%    ~     (p=0.690 n=5+5)
ListenerGeneration/telemetry-12              12.2ms ±16%        10.9ms ± 5%    ~     (p=0.151 n=5+5)
ListenerGeneration/virtualservice-12          895µs ± 9%         840µs ± 5%    ~     (p=0.095 n=5+5)

name                                   old kb/msg         new kb/msg         delta
ListenerGeneration/gateways-12                  875 ± 0%           875 ± 0%    ~     (all equal)
ListenerGeneration/gateways-shared-12          1.77 ± 0%          1.77 ± 0%    ~     (all equal)
ListenerGeneration/empty-12                    9.53 ± 0%          9.53 ± 0%    ~     (all equal)
ListenerGeneration/telemetry-12                19.4 ± 0%          19.4 ± 0%    ~     (all equal)
ListenerGeneration/virtualservice-12           3.86 ± 0%          3.86 ± 0%    ~     (all equal)

name                                   old resources/msg  new resources/msg  delta
ListenerGeneration/gateways-12                 2.00 ± 0%          2.00 ± 0%    ~     (all equal)
ListenerGeneration/gateways-shared-12          2.00 ± 0%          2.00 ± 0%    ~     (all equal)
ListenerGeneration/empty-12                    6.00 ± 0%          6.00 ± 0%    ~     (all equal)
ListenerGeneration/telemetry-12                6.00 ± 0%          6.00 ± 0%    ~     (all equal)
ListenerGeneration/virtualservice-12           3.00 ± 0%          3.00 ± 0%    ~     (all equal)

name                                   old alloc/op       new alloc/op       delta
ListenerGeneration/gateways-12               5.21MB ± 0%        5.13MB ± 0%  -1.54%  (p=0.008 n=5+5)
ListenerGeneration/gateways-shared-12        11.4kB ± 0%        11.3kB ± 0%  -0.68%  (p=0.008 n=5+5)
ListenerGeneration/empty-12                   778kB ± 0%         778kB ± 0%    ~     (p=1.000 n=5+5)
ListenerGeneration/telemetry-12               840kB ± 0%         839kB ± 0%    ~     (p=0.310 n=5+5)
ListenerGeneration/virtualservice-12         55.5kB ± 0%        55.5kB ± 0%    ~     (all equal)

name                                   old allocs/op      new allocs/op      delta
ListenerGeneration/gateways-12                59.1k ± 0%         57.1k ± 0%  -3.38%  (p=0.008 n=5+5)
ListenerGeneration/gateways-shared-12           160 ± 0%           158 ± 0%  -1.25%  (p=0.008 n=5+5)
ListenerGeneration/empty-12                   14.6k ± 0%         14.6k ± 0%    ~     (p=1.000 n=5+5)
ListenerGeneration/telemetry-12               15.3k ± 0%         15.3k ± 0%    ~     (p=0.302 n=5+5)
ListenerGeneration/virtualservice-12            780 ± 0%           780 ± 0%    ~     (all equal)
```